+++ LogicAppUnit/ITestRunner.cs
@@ -164,8 +164,36 @@ namespace LogicAppUnit
         /// <returns>The response from the workflow.</returns>
         /// <remarks>
         /// An empty request body may be used for workflows that contain triggers that do not use a request body, for example a Recurrence trigger.
+        /// This method waits for the workflow to complete before returning a response.
         /// </remarks>
         HttpResponseMessage TriggerWorkflow(HttpMethod method, Dictionary<string, string> requestHeaders = null);
+
+        /// <summary>
+        /// Trigger a workflow using query parameters, an empty request content and optional request headers.
+        /// </summary>
+        /// <param name="queryParams">The query parameters to be passed to the workflow.</param>
+        /// <param name="method">The HTTP method, this needs to match the method defined in the HTTP trigger in the workflow.</param>
+        /// <param name="requestHeaders">The request headers.</param>
+        /// <returns>The response from the workflow.</returns>
+        /// <remarks>
+        /// An empty request body may be used for workflows that contain triggers that do not use a request body, for example a Recurrence trigger.
+        /// This method waits for the workflow to complete before returning a response.
+        /// </remarks>
+        HttpResponseMessage TriggerWorkflow(Dictionary<string, string> queryParams, HttpMethod method, Dictionary<string, string> requestHeaders = null);
+
+        /// <summary>
+        /// Trigger a workflow using query parameters, a relative path and optional request headers.
+        /// </summary>
+        /// <param name="queryParams">The query parameters to be passed to the workflow.</param>
+        /// <param name="method">The HTTP method, this needs to match the method defined in the HTTP trigger in the workflow.</param>
+        /// <param name="relativePath">The relative path to be used in the trigger. The path must already be URL-encoded.</param>
+        /// <param name="requestHeaders">The request headers.</param>
+        /// <returns>The response from the workflow.</returns>
+        /// <remarks>
+        /// An empty request body may be used for workflows that contain triggers that do not use a request body, for example a Recurrence trigger.
+        /// This method waits for the workflow to complete before returning a response.
+        /// </remarks>
+        HttpResponseMessage TriggerWorkflow(Dictionary<string, string> queryParams, HttpMethod method, string relativePath, Dictionary<string, string> requestHeaders = null);
 
         /// <summary>
         /// Trigger a workflow using a request body and optional request headers.
@@ -174,6 +202,9 @@ namespace LogicAppUnit
         /// <param name="method">The HTTP method, this needs to match the method defined in the HTTP trigger in the workflow.</param>
         /// <param name="requestHeaders">The request headers.</param>
         /// <returns>The response from the workflow.</returns>
+        /// <remarks>
+        /// This method waits for the workflow to complete before returning a response.
+        /// </remarks>
         HttpResponseMessage TriggerWorkflow(HttpContent content, HttpMethod method, Dictionary<string, string> requestHeaders = null);
 
         /// <summary>
@@ -184,9 +215,42 @@ namespace LogicAppUnit
         /// <param name="relativePath">The relative path to be used in the trigger. The path must already be URL-encoded.</param>
         /// <param name="requestHeaders">The request headers.</param>
         /// <returns>The response from the workflow.</returns>
+        /// <remarks>
+        /// This method waits for the workflow to complete before returning a response.
+        /// </remarks>
         HttpResponseMessage TriggerWorkflow(HttpContent content, HttpMethod method, string relativePath, Dictionary<string, string> requestHeaders = null);
 
+        /// <summary>
+        /// Trigger a workflow using query parameters, a request body, a relative path and optional request headers.
+        /// </summary>
+        /// <param name="queryParams">The query parameters to be passed to the workflow.</param>
+        /// <param name="content">The content (including any content headers) for running the workflow, or <c>null</c> if there is no content.</param>
+        /// <param name="method">The HTTP method, this needs to match the method defined in the HTTP trigger in the workflow.</param>
+        /// <param name="relativePath">The relative path to be used in the trigger. The path must already be URL-encoded.</param>
+        /// <param name="requestHeaders">The request headers.</param>
+        /// <returns>The response from the workflow.</returns>
+        /// <remarks>
+        /// This method waits for the workflow to complete before returning a response.
+        /// </remarks>
+        HttpResponseMessage TriggerWorkflow(Dictionary<string, string> queryParams, HttpContent content, HttpMethod method, string relativePath, Dictionary<string, string> requestHeaders = null);
+
         #endregion // TriggerWorkflow
+
+        #region WaitForAsynchronousResponse
+
+        /// <summary>
+        /// Configure the test runner (in seconds) to wait for and return the asynchronous response.
+        /// </summary>
+        /// <param name="maxTimeoutSeconds">The maximum number of seconds to poll for the asynchronous response after which test runner will time out.</param>
+        void WaitForAsynchronousResponse(int maxTimeoutSeconds);
+
+        /// <summary>
+        /// Configure the test runner to wait for and return the asynchronous response.
+        /// </summary>
+        /// <param name="maxTimeout">The maximum time to poll for the asynchronous response after which test runner will time out.</param>
+        void WaitForAsynchronousResponse(TimeSpan maxTimeout);
+
+        #endregion // WaitForAsynchronousResponse
 
         /// <summary>
         /// Wraps a test assertion in a <c>catch</c> block which logs additional workflow execution information when the assertion fails.