+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# 1.6.0 (12th June 2023)
+# 1.7.0 (24th July 2023)
 
 LogicAppUnit Testing Framework:
 
@@ -8,6 +8,25 @@ LogicAppUnit Testing Framework:
 LogicAppUnit.Samples.LogicApps.Tests:
 
 - Added a `fluent-workflow` workflow and unit tests to demonstrate the use of the fluent API.
+
+
+# 1.6.0 (5th June 2023)
+
+LogicAppUnit Testing Framework:
+
+- Added support for Linux and MacOS platforms. [[PR #14](https://github.com/LogicAppUnit/TestingFramework/pull/14), [@easchi](https://github.com/eashi) and [PR #15](https://github.com/LogicAppUnit/TestingFramework/pull/15), [@atthevergeof](https://github.com/atthevergeof)]
+- Added new overloads to `TestRunner.TriggerWorkflow()` to configure URL query parameters when triggering a workflow with a HTTP trigger. [[PR #15](https://github.com/LogicAppUnit/TestingFramework/pull/15), [@atthevergeof](https://github.com/atthevergeof)]
+- Added support for workflows using a HTTP trigger with an asynchronous response. Previous versions of the framework assumed that all responses were synchronous. Now the framework handles a 202 (Accepted) response and uses the callback URL (in the `Location` header) to poll the workflow and get the response. [[PR #15](https://github.com/LogicAppUnit/TestingFramework/pull/15), [@atthevergeof](https://github.com/atthevergeof)]
+- Added support for workflows using a HTTP trigger where the response action is not the last action in the workflow. Previous versions of the framework assumed that the workflow was complete once the response was received. Now the framework polls the workflow status to ensure that the workflow has completed. [[PR #15](https://github.com/LogicAppUnit/TestingFramework/pull/15), [@atthevergeof](https://github.com/atthevergeof)]
+- Fixed a bug in `TestRunner.TriggerWorkflow()` where the return value was being incorrectly set to the (disposed) workflow run history API response. The response is now correctly set to the workflow trigger API response. This bug only occurred for workflows that have a non-HTTP trigger (which is then replaced by a HTTP trigger by the framework).
+  <br />
+  :warning: ***This is a breaking change. Previously the status code for the response would have been HTTP 200 (OK), now it will be HTTP 202 (Accepted).***
+
+LogicAppUnit.Samples.LogicApps.Tests:
+
+- Added a `http-async-workflow` workflow and unit tests to demonstrate the use of the testing framework with HTTP triggers and asynchronous responses.
+
+Thanks to [@easchi](https://github.com/eashi) and [@atthevergeof](https://github.com/atthevergeof) for their contributions.
 
 
 # 1.5.0 (14th April 2023)